# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daily Task Manager API for incidents, service requests, work notes, and resolutions
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-requests:
    get:
      operationId: listServiceRequests
      tags: [service-requests]
      summary: List all service requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of service requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createServiceRequest
      tags: [service-requests]
      summary: Create a new service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /service-requests/{id}:
    get:
      operationId: getServiceRequest
      tags: [service-requests]
      summary: Get a service request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateServiceRequest
      tags: [service-requests]
      summary: Update a service request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestUpdate"
      responses:
        "200":
          description: Updated service request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteServiceRequest
      tags: [service-requests]
      summary: Delete a service request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Work Notes ───────────────────────────────────────────────
  /work-notes:
    get:
      operationId: listWorkNotes
      tags: [work-notes]
      summary: List all work notes
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: relatedType
          in: query
          schema:
            type: string
            enum: [incident, service_request, general]
        - name: relatedId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of work notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkNote"
    post:
      operationId: createWorkNote
      tags: [work-notes]
      summary: Create a new work note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkNote"
  /work-notes/{id}:
    get:
      operationId: getWorkNote
      tags: [work-notes]
      summary: Get a work note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work note details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkNote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWorkNote
      tags: [work-notes]
      summary: Update a work note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkNoteUpdate"
      responses:
        "200":
          description: Updated work note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkNote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWorkNote
      tags: [work-notes]
      summary: Delete a work note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Resolutions ──────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Incident ──────────────────────────────────────────────
...
    ServiceRequest:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        requester:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        requestDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServiceRequestInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        requester:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
        tags:
          type: string
        requestDate:
          type: string
          format: date
    ServiceRequestUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        requester:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
        tags:
          type: string
        requestDate:
          type: string
          format: date

    # ─── Work Note ─────────────────────────────────────────────
    WorkNote:
      type: object
      required: [id, title, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        relatedType:
          type: ["string", "null"]
          enum: [incident, service_request, general, null]
        relatedId:
          type: ["integer", "null"]
        noteDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkNoteInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        relatedType:
          type: string
          enum: [incident, service_request, general]
        relatedId:
          type: integer
        noteDate:
          type: string
          format: date
    WorkNoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        relatedType:
          type: string
          enum: [incident, service_request, general]
        relatedId:
          type: integer
        noteDate:
          type: string
          format: date

    # ─── Resolution ────────────────────────────────────────────
...